Define is_async function at runtime init
